Move sanitizers to use ModifyingDataVisitor
