+++ src/main/java/org/parchmentmc/compass/data/sanitation/impl/BouncerDataMover.java
@@ -1,11 +1,13 @@
 package org.parchmentmc.compass.data.sanitation.impl;
 
 import org.checkerframework.checker.nullness.qual.Nullable;
-import org.parchmentmc.compass.data.sanitation.AbstractSanitizer;
+import org.parchmentmc.compass.data.sanitation.Sanitizer;
+import org.parchmentmc.feather.mapping.MappingDataContainer;
 import org.parchmentmc.feather.metadata.BouncingTargetMetadata;
 import org.parchmentmc.feather.metadata.ClassMetadata;
 import org.parchmentmc.feather.metadata.MethodMetadata;
 import org.parchmentmc.feather.metadata.Reference;
+import org.parchmentmc.feather.metadata.SourceMetadata;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -16,7 +18,7 @@ import static org.parchmentmc.feather.mapping.MappingDataContainer.MethodData;
 /**
  * Moves data in bouncer methods to their targets.
  */
-public class BouncerDataMover extends AbstractSanitizer {
+public class BouncerDataMover extends Sanitizer {
     // First pass is to collect bouncers and delete them, second pass is applying the data from bouncers to their targets
     private boolean finishedCollecting = false;
     private final Map<String, MethodData> data = new HashMap<>();
@@ -26,12 +28,12 @@ public class BouncerDataMover extends AbstractSanitizer {
     }
 
     @Override
-    public boolean start(boolean isMetadataAvailable) {
-        return isMetadataAvailable; // Skip if metadata is not available
+    public boolean visit(MappingDataContainer container, @Nullable SourceMetadata metadata) {
+        return metadata != null; // Skip if metadata is not available
     }
 
     @Override
-    public Action<MethodData> sanitize(ClassData classData, MethodData methodData,
+    public Action<MethodData> modifyMethod(ClassData classData, MethodData methodData,
                                        @Nullable ClassMetadata classMetadata, @Nullable MethodMetadata methodMetadata) {
         if (!finishedCollecting && methodMetadata != null && methodMetadata.getBouncingTarget().isPresent()) {
             final BouncingTargetMetadata targetData = methodMetadata.getBouncingTarget().get();
