+++ src/main/java/org/parchmentmc/compass/data/sanitation/impl/DescriptorParametersSanitizer.java
@@ -2,16 +2,20 @@ package org.parchmentmc.compass.data.sanitation.impl;
 
 import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
 import org.checkerframework.checker.nullness.qual.Nullable;
-import org.parchmentmc.compass.data.sanitation.AbstractSanitizer;
+import org.parchmentmc.compass.data.sanitation.Sanitizer;
 import org.parchmentmc.compass.util.DescriptorIndexer;
+import org.parchmentmc.feather.mapping.MappingDataContainer;
 import org.parchmentmc.feather.metadata.ClassMetadata;
 import org.parchmentmc.feather.metadata.MethodMetadata;
+import org.parchmentmc.feather.metadata.SourceMetadata;
 
 import java.util.BitSet;
 
-import static org.parchmentmc.feather.mapping.MappingDataContainer.*;
+import static org.parchmentmc.feather.mapping.MappingDataContainer.ClassData;
+import static org.parchmentmc.feather.mapping.MappingDataContainer.MethodData;
+import static org.parchmentmc.feather.mapping.MappingDataContainer.ParameterData;
 
-public class DescriptorParametersSanitizer extends AbstractSanitizer {
+public class DescriptorParametersSanitizer extends Sanitizer {
     public DescriptorParametersSanitizer() {
         super("descriptor parameter indexes");
     }
@@ -20,14 +24,14 @@ public class DescriptorParametersSanitizer extends AbstractSanitizer {
     private DescriptorIndexer indexer = null;
 
     @Override
-    public boolean start(boolean isMetadataAvailable) {
+    public boolean visit(MappingDataContainer container, @Nullable SourceMetadata metadata) {
         indexer = new DescriptorIndexer();
-        return super.start(isMetadataAvailable);
+        return super.visit(container, metadata);
     }
 
     @Override
-    public Action<ParameterData> sanitize(ClassData classData, MethodData methodData, ParameterData paramData,
-                                          @Nullable ClassMetadata classMetadata, @Nullable MethodMetadata methodMetadata) {
+    public Action<ParameterData> modifyParameter(ClassData classData, MethodData methodData, ParameterData paramData,
+                                                 @Nullable ClassMetadata classMetadata, @Nullable MethodMetadata methodMetadata) {
         final BitSet indexes = indexer.getIndexes(methodData, methodMetadata);
 
         if (!indexes.get(paramData.getIndex())) {
